fix(tickets): isolate mention geometry helpers
Move mention layout calculations out of the component to satisfy hot-reload linting.
